added to README boot on hardware guide, integrated frambuffer with main code

+++ include/stdint.h
@@ -12,7 +12,7 @@ typedef long long int int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
+typedef unsigned long int uint64_t;
 
 // Minimum-width integer types
 typedef int8_t int_least8_t;
@@ -44,14 +44,14 @@ typedef uint64_t uintptr_t;
 typedef int64_t intmax_t;
 typedef uint64_t uintmax_t;
 
-typedef unsigned long int size_t;
-typedef long int ssize_t;
+typedef long unsigned int size_t;
+typedef int64_t ssize_t;
 
 // Limits of exact-width integer types
 #define INT8_MIN (-128)
 #define INT16_MIN (-32768)
-#define INT32_MIN (-2147483647-1)
-#define INT64_MIN (-9223372036854775807LL-1)
+#define INT32_MIN (-2147483647 - 1)
+#define INT64_MIN (-9223372036854775807LL - 1)
 
 #define INT8_MAX 127
 #define INT16_MAX 32767
@@ -140,4 +140,4 @@ typedef long int ssize_t;
 #define PRIx64 "llx"
 #define PRIX64 "llX"
 
-#endif // _STDINT_H_
+#endif  // _STDINT_H_
